added files and ranks to notation
