+++ notation.py
@@ -22,3 +22,23 @@ def get_piece_by_character(character: str):
         raise ValueError(character)
 
     return Piece(color, piece_type)
+
+
+def _get_char_number_in_alphabet(char):
+    return ord(char) - 97
+
+
+def get_file_by_char(char):
+    num_in_alphabet = _get_char_number_in_alphabet(char.lower())
+
+    if num_in_alphabet < 0 or num_in_alphabet > 7:
+        raise ValueError(char)
+
+    return num_in_alphabet
+
+
+def get_rank_by_char(char):
+    try:
+        return int(char) - 1
+    except ValueError:
+        raise ValueError(char)
